feat: link category to user

+++ src/house-resource-manager/presentation/routers/category.router.ts
@@ -1,41 +1,57 @@
+import { getIdFromRequest } from "core.utils";
 import { Router } from "express";
 import { BaseDomainError, DomainErrorType } from "../../domain/errors";
 import { categoryUseCasesService } from "../../infra/di";
+import { UnauthorizedError } from "../errors";
 import { CategoryValidator } from "../validators";
 
 const categoryRouter = Router();
 
-categoryRouter.get("/", async (_req, res) => {
-	const result = await categoryUseCasesService.getResources();
+categoryRouter.get("/", async (req, res) => {
+	const userId = getIdFromRequest(req);
+	if (!userId) throw new UnauthorizedError();
+
+	const result = await categoryUseCasesService.getResources({ userId });
 	res.status(200).json(result);
 });
 
 categoryRouter.post("/", async (req, res) => {
+	const userId = getIdFromRequest(req);
+	if (!userId) throw new UnauthorizedError();
+
 	const input = CategoryValidator.createValidator.parse(req.body);
-	const result = await categoryUseCasesService.createResource(input);
+	const result = await categoryUseCasesService.createResource({
+		...input,
+		userId,
+	});
 	res.status(201).json(result);
 });
 
 categoryRouter.put("/:id", async (req, res) => {
 	const id = req.params.id;
-
 	if (!id) throw idFieldNotFoundError;
 
+	const userId = getIdFromRequest(req);
+	if (!userId) throw new UnauthorizedError();
+
 	const input = CategoryValidator.updateValidator.parse(req.body);
-	const result = await categoryUseCasesService.updateResource({
+	const result = await categoryUseCasesService.updateCategory({
 		id,
 		name: input.name,
 		description: input.description,
+		userId,
 	});
 	res.status(200).json(result);
 });
 
 categoryRouter.delete("/:id", async (req, res) => {
 	const id = req.params.id;
-
 	if (!id) throw idFieldNotFoundError;
 
-	await categoryUseCasesService.deleteResource({ id });
+	const userId = getIdFromRequest(req);
+	if (!userId) throw new UnauthorizedError();
+
+	await categoryUseCasesService.deleteResource({ id, userId });
 	res.status(204).json();
 });
